+++ context/app/api/vitessce_confs/test_confs.py
@@ -17,7 +17,7 @@ from .assay_confs import (
     StitchedCytokitSPRMViewConfBuilder
 )
 
-MOCK_NEXUS_TOKEN = "nexus_token"
+MOCK_GROUPS_TOKEN = "groups_token"
 
 FIXTURES_INPUT_DIR = "fixtures/input_entity"
 FIXTURES_EXPECTED_OUTPUT_DIR = "fixtures/output_expected"
@@ -33,6 +33,7 @@ AssayConfClasses = {
     "seqfish": SeqFISHViewConfBuilder,
     "sprm": SPRMJSONViewConfBuilder,
     "rna_zarr": RNASeqAnnDataZarrViewConfBuilder,
+    "rna_azimuth_zarr": RNASeqAnnDataZarrViewConfBuilder,
     "spatial_rna_zarr": SpatialRNASeqAnnDataZarrViewConfBuilder,
     "sprm_anndata": SPRMAnnDataViewConfBuilder,
     "cytokit_anndata": StitchedCytokitSPRMViewConfBuilder
@@ -55,7 +56,7 @@ def test_assays():
         if assay in ["sprm", "sprm_anndata"]:
             vc = AssayViewConfClass(
                 entity=entity,
-                nexus_token=MOCK_NEXUS_TOKEN,
+                groups_token=MOCK_GROUPS_TOKEN,
                 is_mock=True,
                 base_name=BASE_NAME_FOR_SPRM,
                 imaging_path="imaging_path",
@@ -65,7 +66,7 @@ def test_assays():
             )
         else:
             vc = AssayViewConfClass(
-                entity=entity, nexus_token=MOCK_NEXUS_TOKEN, is_mock=True
+                entity=entity, groups_token=MOCK_GROUPS_TOKEN, is_mock=True
             )
         conf = vc.get_conf_cells().conf
         conf_expected = json.loads(
@@ -81,7 +82,7 @@ def test_assays():
         if assay in ["sprm", "sprm_anndata"]:
             vc = AssayViewConfClass(
                 entity=malformed_entity,
-                nexus_token=MOCK_NEXUS_TOKEN,
+                groups_token=MOCK_GROUPS_TOKEN,
                 is_mock=True,
                 base_name=BASE_NAME_FOR_SPRM,
                 imaging_path="imaging_path",
@@ -91,7 +92,7 @@ def test_assays():
             )
         else:
             vc = AssayViewConfClass(
-                entity=malformed_entity, nexus_token=MOCK_NEXUS_TOKEN, is_mock=True
+                entity=malformed_entity, groups_token=MOCK_GROUPS_TOKEN, is_mock=True
             )
         with pytest.raises(FileNotFoundError):
             vc.get_conf_cells().conf