+++ context/app/static/js/components/organ/Assays/Assays.jsx
@@ -6,7 +6,6 @@ import Paper from '@material-ui/core/Paper';
 import Button from '@material-ui/core/Button';
 
 import { SecondaryBackgroundTooltip } from 'js/shared-styles/tooltips';
-import { getDefaultQuery } from 'js/helpers/functions';
 import EntitiesTable from 'js/shared-styles/tables/EntitiesTable';
 import { LightBlueLink } from 'js/shared-styles/Links';
 import SectionHeader from 'js/shared-styles/sections/SectionHeader';
@@ -14,7 +13,7 @@ import SectionContainer from 'js/shared-styles/sections/SectionContainer';
 import { SpacedSectionButtonRow } from 'js/shared-styles/sections/SectionButtonRow';
 import useSearchData from 'js/hooks/useSearchData';
 
-import { Flex, StyledInfoIcon } from '../style';
+import { Flex, StyledInfoIcon, StyledDatasetIcon } from '../style';
 import { getSearchURL } from '../utils';
 
 function Assays(props) {
@@ -24,7 +23,6 @@ function Assays(props) {
   const query = useMemo(
     () => ({
       size: 0,
-      query: getDefaultQuery(),
       aggs: {
         mapped_data_types: {
           filter: {
@@ -73,6 +71,7 @@ function Assays(props) {
         }
         buttons={
           <Button color="primary" variant="contained" component="a" href={searchUrl}>
+            <StyledDatasetIcon />
             View All Datasets
           </Button>
         }