+++ context/app/static/js/components/organ/OrganDatasetsChart/index.js
@@ -1,0 +1,3 @@
+import OrganDatasetsChart from './OrganDatasetsChart';
+
+export default OrganDatasetsChart;