+++ context/app/static/js/components/searchPage/filters/CheckboxFilterItem/index.js
@@ -1,0 +1,3 @@
+import CheckboxFilterItem from './CheckboxFilterItem';
+
+export default CheckboxFilterItem;