+++ context/app/static/js/components/tutorials/SearchDatasetTutorial/config.js
@@ -2,7 +2,7 @@ const sortTileViewStepTitle = 'Sort Search Results for Tile View';
 
 const defaultSteps = [
   {
-    target: '#Data-Type div.sk-item-list > div:nth-child(1)',
+    target: '#Data-Type div.sk-item-list > label:nth-child(1)',
     disableBeacon: true,
     content:
       'The Dataset Metadata menu on the left side allows filtering datasets by any combination of metadata categories: Data Type, Organ and Specimen Type. Search results update automatically as you edit the selection of filters.',