+++ context/app/static/js/pages/Organ/Organ.jsx
@@ -10,6 +10,7 @@ import Assays from 'js/components/organ/Assays';
 import Description from 'js/components/organ/Description';
 import OrganInfo from 'js/components/organ/OrganInfo';
 import Samples from 'js/components/organ/Samples';
+import DatasetsBarChart from 'js/components/organ/OrganDatasetsChart';
 
 import { FlexRow, Content } from './style';
 
@@ -58,6 +59,7 @@ function Organ(props) {
         {shouldDisplaySection[searchId] && (
           <div id={searchId}>
             <Assays searchTerms={organ.search} />
+            <DatasetsBarChart name={organ.name} search={organ.search} />
             <Samples searchTerms={organ.search} />
           </div>
         )}