+++ context/app/static/js/pages/Publication/Publication.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Typography from '@material-ui/core/Typography';
 
 import Markdown from 'js/components/Markdown';
-import VisualizationWrapper from 'js/components/Detail/visualization/VisualizationWrapper';
+import VisualizationWrapper from 'js/components/detailPage/visualization/VisualizationWrapper';
 import SectionHeader from 'js/shared-styles/sections/SectionHeader';
 import { StyledPaper } from './style';
 
@@ -35,9 +35,9 @@ function Publication(props) {
         </Typography>
         <b>Corresponding Author:</b>{' '}
         {authors.corresponding.map((author) => (
-          <>
+          <span key={author.name}>
             {author.name} - <a href={`mailto:${author.email}`}>{author.email}</a>
-          </>
+          </span>
         ))}
       </StyledPaper>
       {Boolean(vitessce_conf) && <VisualizationWrapper vitData={vitessce_conf} />}