+++ context/app/static/js/shared-styles/charts/StackedBar/style.js
@@ -1,0 +1,13 @@
+import styled, { css } from 'styled-components';
+
+const StyledRect = styled.rect`
+  ${(props) =>
+    props.$showHover &&
+    css`
+      &:hover {
+        filter: brightness(50%);
+      }
+    `}
+`;
+
+export { StyledRect };