+++ context/build-utils/alias.js
@@ -3,10 +3,12 @@ const { resolve } = require('path');
 
 const alias = {
   'metadata-field-descriptions$': resolve('./ingest-validation-tools/docs/field-descriptions.yaml'),
+  'metadata-field-types$': resolve('./ingest-validation-tools/docs/field-types.yaml'),
   js: resolve(__dirname, '../app/static/js/'),
   'portal-images': resolve(__dirname, '../app/static/portal-images/'),
   'shared-styles': resolve(__dirname, '../app/static/js/shared-styles/'),
   package: resolve(__dirname, '../package.json'),
+  'txml/txml': 'txml/dist/txml'
 };
 
 exports.alias = alias;